Add test for float16 support with conv1d and update docstring (#16273)
* Add test for float16 support with conv1d and update docs

The float16 support for conv1d has already been in place.
However there was no test for float16 with conv1d.
This fix adds test case to cover float16 support with conv1d.

In addition, float64 support with conv1d is not possible
because conv1d calls conv2d, which in turn does not support
float64 yet (See 12941, 13097, and 12943)
The docstring incorrectly claims float64 support with conv1d.

This fix updates related docstring to remove float64 part.


* Update docstring of conv1d to remove float64 and add float16.


* Update docstring with TODO.

+++ tensorflow/python/ops/nn_ops.py
@@ -2303,7 +2303,7 @@ def conv1d(value, filters, stride, padding,
   returned to the caller.
 
   Args:
-    value: A 3D `Tensor`.  Must be of type `float32` or `float64`.
+    value: A 3D `Tensor`.  Must be of type `float16` or `float32`.
     filters: A 3D `Tensor`.  Must have the same type as `input`.
     stride: An `integer`.  The number of entries by which
       the filter is moved right at each step.
